+++ apps/web/src/modules/common/lib/apollo/apollo-wrapper.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import { config } from "@/config/constants";
+import { HttpLink } from "@apollo/client";
+import {
+    ApolloNextAppProvider,
+    ApolloClient,
+    InMemoryCache,
+} from "@apollo/experimental-nextjs-app-support";
+
+const { API_URL } = config;
+
+function makeClient() {
+    const httpLink = new HttpLink({
+        uri: API_URL,
+        fetchOptions: { cache: "force-cache" },
+    });
+    return new ApolloClient({
+        cache: new InMemoryCache(),
+        link: httpLink,
+    });
+}
+
+export function ApolloWrapper({ children }: React.PropsWithChildren) {
+    return (
+        <ApolloNextAppProvider makeClient={makeClient}>
+            {children}
+        </ApolloNextAppProvider>
+    );
+}