+++ apps/web/src/modules/docs/components/clients-query-test.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import { useSuspenseQuery } from "@apollo/client";
+
+import {
+    GetServicePageInfoDocument,
+    GetServicesInfoQuery,
+} from "@/graphql/generated/graphql";
+
+const ClientQueryTest = () => {
+    const { data, error } = useSuspenseQuery<GetServicesInfoQuery>(
+        GetServicePageInfoDocument,
+    );
+
+    return (
+        <div>
+            {data.services &&
+                data.services.edges.map(({ node: post }) => {
+                    return <div key={post.slug}>{post.title}</div>;
+                })}
+        </div>
+    );
+};
+
+export default ClientQueryTest;