+++ apps/web/src/modules/services/utils/index.ts
@@ -1,75 +1,75 @@
-export const mapServicesPage = (page: any, services: any) => {
-    const mappedPage = {
-        title: page.title,
-        servicesSection: {
-            title: page.title,
-            description: page.servicesPage.description,
-            items: services.map((service: any) => mapServicePage(service.node)),
-            button: {
-                text: page.servicesPage.button.link.title,
-                link: page.servicesPage.button.link.url,
-                variant: page.servicesPage.button.variant,
-                size: page.servicesPage.button.size,
-                icon: {
-                    name: page.servicesPage.button.icon.name,
-                    size: page.servicesPage.button.icon.size,
-                },
-            },
-            className: page.servicesPage.className,
-        },
-        letsHelp: {
-            title: page.letsHelp?.title,
-            description: page.letsHelp?.description,
-            className: page.letsHelp?.className,
-            img: {
-                alt: page.letsHelp?.img.node.altText,
-                src: page.letsHelp?.img.node.sourceUrl,
-            },
-            button: {
-                text: page.letsHelp?.button.text,
-                variant: page.letsHelp?.button.variant,
-                link: page.letsHelp?.button.link.url,
-                size: page.letsHelp?.button.size,
-            },
-        },
-    };
+import { ServiceSectionProps } from "@/modules/services/components/services-section";
+import { ClinicSectionProps } from "../components/clinic-section";
+import {
+    GetServicePageInfoQuery,
+    GetServiceBySlugQuery,
+} from "@/graphql/generated/graphql";
 
-    return mappedPage;
+import {
+    getServices,
+    ServicePageInfo,
+} from "@/modules/services/actions/services.action";
+
+import { getCallToAction } from "@/modules/common/actions/callToActions.action";
+
+export const mapServicePageData = async (
+    data: GetServicePageInfoQuery["page"],
+): Promise<ServiceSectionProps> => {
+    const serviceData = data?.servicePageSettings?.servicesSection;
+
+    return {
+        title: serviceData?.title || "",
+        description: serviceData?.description || "",
+        className: serviceData?.className || "",
+        items: await getServices(),
+    };
 };
 
-export const mapServicePage = (page: any) => {
-    if (!page) return null;
+export const mapClinicSectionData = async (
+    data: GetServicePageInfoQuery["page"],
+): Promise<ClinicSectionProps> => {
+    const clinicData = data?.servicePageSettings?.clinicSection;
 
-    const mappedPage = {
-        title: page.title,
-        description: page.servicePage.description,
-        button: {
-            text: page.servicePage.button.link.title,
-            link: page.servicePage.button.link.url,
-            variant: page.servicePage.button.variant,
-            size: page.servicePage.button.size,
-        },
-        icon: page.servicePage.button.icon.name,
-        iconsize: page.servicePage.button.icon.size,
-        className: page.servicePage.className,
-        slug: page.slug,
-        content: page.content,
-        letsHelp: {
-            title: page.letsHelp?.title,
-            description: page.letsHelp?.description,
-            className: page.letsHelp?.className,
-            img: {
-                alt: page.letsHelp?.img?.node?.altText,
-                src: page.letsHelp?.img?.node?.sourceUrl,
-            },
-            button: {
-                text: page.letsHelp?.button?.text,
-                variant: page.letsHelp?.button?.variant,
-                link: page.letsHelp?.button?.link?.url,
-                size: page.letsHelp?.button?.size,
-            },
-        },
+    const clinicSection: ClinicSectionProps = {
+        title: clinicData?.title || "",
+        items:
+            clinicData?.imagesClinic?.edges?.map(({ node: item }) => ({
+                src: item?.sourceUrl || "",
+                alt: item?.altText || "",
+            })) || [],
     };
 
-    return mappedPage;
+    return clinicSection;
+};
+
+export const mapCalltoActionsServices = async (
+    data: GetServicePageInfoQuery["page"],
+): Promise<ServicePageInfo["callToActions"]> => {
+    return await Promise.all(
+        (data?.servicePageSettings?.callToActions?.edges || []).map(
+            async ({ node: cta }) => await getCallToAction(cta.id),
+        ),
+    );
+};
+
+export const mapCalltoActionsService = async (
+    data: GetServiceBySlugQuery["service"],
+): Promise<ServicePageInfo["callToActions"]> => {
+    return await Promise.all(
+        (data?.serviceSettings?.callToActions?.edges || []).map(
+            async ({ node: cta }) => await getCallToAction(cta.id),
+        ),
+    );
+};
+
+export const mapCallusActionsService = async (
+    data: GetServiceBySlugQuery["service"],
+): Promise<ServicePageInfo["callToActions"]> => {
+    const result = await Promise.all(
+        (data?.serviceSettings?.callUsAction?.edges || []).map(
+            async ({ node: cta }) => await getCallToAction(cta.id),
+        ),
+    );
+
+    return result;
 };