+++ packages/ui/tailwind.config.ts
@@ -31,7 +31,7 @@ const config = {
                 link: "var(--link)",
                 input: "var(--input)",
                 "dark-primary": "var(--dark-primary)",
-                "dark-grey": "var(--dark-grey)",
+                "dark-grey": "hsl(var(--dark-grey))",
                 primary: {
                     DEFAULT: "hsl(var(--primary))",
                     foreground: "var(--primary-foreground)",
@@ -105,92 +105,7 @@ const config = {
             },
         },
     },
-    extend: {
-      screens: {
-        "xmd": "977px",
-      },
-      colors: {
-        border: "hsl(var(--border))",
-        ring: "hsl(var(--ring))",
-        background: "var(--background)",
-        foreground: "var(--foreground)",
-        link: "var(--link)",
-        input: "var(--input)",
-        "dark-primary": "var(--dark-primary)",
-        "dark-grey": "hsl(var(--dark-grey))",
-        primary: {
-          DEFAULT: "hsl(var(--primary))",
-          foreground: "var(--primary-foreground)",
-        },
-        secondary: {
-          DEFAULT: "hsl(var(--secondary))",
-          foreground: "var(--secondary-foreground)",
-        },
-        dark: {
-            DEFAULT: "var(--dark)",
-            foreground: "var(--dark-foreground)",
-        },
-        light: {
-          DEFAULT: "var(--light)",
-          foreground: "var(--light-foreground)",
-        },
-        info: {
-            DEFAULT: "var(--info)",
-            foreground: "var(--info-foreground)",
-        },
-        "dark-info": {
-            DEFAULT: "var(--dark-info)",
-            foreground: "var(--dark-info-foreground)",
-        },
-        "light-info": {
-            DEFAULT: "var(--light-info)",
-            foreground: "var(--light-info-foreground)",
-        },
-        destructive: {
-          DEFAULT: "hsl(var(--destructive))",
-          foreground: "hsl(var(--destructive-foreground))",
-        },
-        muted: {
-          DEFAULT: "var(--muted)",
-          foreground: "var(--muted-foreground)",
-        },
-        accent: {
-          DEFAULT: "hsl(var(--accent))",
-          foreground: "hsl(var(--accent-foreground))",
-        },
-        popover: {
-          DEFAULT: "hsl(var(--popover))",
-          foreground: "hsl(var(--popover-foreground))",
-        },
-        card: {
-          DEFAULT: "hsl(var(--card))",
-          foreground: "hsl(var(--card-foreground))",
-        },
-      },
-      borderRadius: {
-        lg: "var(--radius)",
-        md: "calc(var(--radius) - 2px)",
-        sm: "calc(var(--radius) - 4px)",
-      },
-      keyframes: {
-        "accordion-down": {
-          from: { height: "0" },
-          to: { height: "var(--radix-accordion-content-height)" },
-        },
-        "accordion-up": {
-          from: { height: "var(--radix-accordion-content-height)" },
-          to: { height: "0" },
-        },
-      },
-      fontFamily: {
-        "noto-sans": "var(--font-noto-sans)",
-      },
-      animation: {
-        "accordion-down": "accordion-down 0.2s ease-out",
-        "accordion-up": "accordion-up 0.2s ease-out",
-      },
-    },
-  plugins: [tailwindcssAnimate],
+    plugins: [tailwindcssAnimate],
 } satisfies Config;
 
 export default config;